some mods on login, register and make user seeder

+++ database/migrations/2019_05_28_174627_create_accounts_table.php
@@ -18,7 +18,7 @@ class CreateAccountsTable extends Migration
             $table->string('type')->default('normal');
             $table->decimal('price', 5, 2)->default(.65);
             $table->bigInteger('message_limit')->default(0);
-            $table->decimal('balance', 5, 2)->default(0);
+            $table->float('balance', 20, 2)->default(0);
             $table->string('status')->default('1');
             $table->timestamps();
             // need FK
